requirements.txt
# ============================================
# 🎩 LA FAMIGLIA LINKS — Dependências Principais
# Flask + IA + Banco + Utilidades
# ============================================

# --- Core do Flask ---
flask==3.0.3
flask-cors==4.0.0
gunicorn==23.0.0

# --- Banco de Dados (SQLite) ---
sqlalchemy==2.0.36

# --- Segurança e JWT ---
pyjwt==2.9.0

# --- Requests e integração com APIs ---
requests==2.32.3

# --- Suporte a imagens e arquivos estáticos ---
pillow==11.0.0

# --- Utilidades gerais ---
python-dotenv==1.0.1

# --- Módulos opcionais de IA (caso use integração local) ---
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
transformers==4.46.2
diffusers==0.31.0
accelerate==1.10.1
safetensors==0.6.2
huggingface-hub==0.35.3

# --- Compatibilidade e desempenho ---
numpy==1.26.4
APScheduler==3.10.4
